encode image src as a URI
This fixes an issue, where files containing spaces could not be
deleted, since the associated img tag can not be found.

In the function onToggleSelected, the path is extracted from
e.target.src, which is a dat URI.
If the file pointed to contained a space in the filename, it will be
encoded as %20.
However, the filename passed to appendImage() does contain the space
and not %20.
Thus the path added to selectedImages is different than the src attribute
of the matching img tag, leading to document.querySelector() returning
null.
Encoding the filename before setting it as img.src fixes this problem.